Separate main function, rename
