+++ svg.py
@@ -1,7 +1,4 @@
 import svgwrite
-import sys
-
-import generate_tray_from_specs as gtray
 
 STROKE = 0.1
 MIN_TOOTH_WIDTH = 8.0
@@ -531,46 +528,5 @@ def draw_edges(dwg, trayspec):
 
     for path in paths:
         dwg.add(path)
-
-
-def finish(dwg):
-    dwg.save()
-
-
-if __name__ == '__main__':
-    if len(sys.argv) != 2:
-        print("Pass the specs file as the parameter")
-        sys.exit(1)
-
-    # TODO: Bring the material specifications from some external source at some point.
-    cfg = {
-        'spacer_width': 3,
-        'edge_width': 3,
-        'spacer_material_width': 400,
-        'spacer_material_height': 300,
-        'edge_material_width': 400,
-        'edge_material_height': 300
-    }
-
-    specs = gtray.get_specification(sys.argv[1])
-
-    trays = gtray.generate_trays_from_spec(cfg['spacer_width'], cfg['edge_width'], specs)
-
-    for tray in trays:
-        for column in tray['columns']:
-            for slot in column['slots']:
-                if 'height' not in slot:
-                    slot['height'] = slot['min-height']
-
-        edge_width = tray['edge_width']
-        spacer_width = tray['spacer_width']
-        tray_name = tray['name']
-        dwg = get_drawing('output/{}-{}-edges.svg'.format(tray_name, edge_width), cfg['edge_material_width'], cfg['edge_material_height'])
-        draw_edges(dwg, tray)
-        finish(dwg)
-
-        dwg = get_drawing('output/{}-{}-spacers.svg'.format(tray_name, spacer_width), cfg['spacer_material_width'], cfg['spacer_material_height'])
-        draw_spacers(dwg, tray)
-        finish(dwg)
